Keep the naive avx512 path on for figuring out codebuild capabilities, turn it off once we know what's supported where.

+++ tests/crc_test.c
@@ -108,9 +108,8 @@ static int s_test_crc32c(struct aws_allocator *allocator, void *ctx) {
     aws_byte_buf_init(&avx_buf, allocator, 512);
     aws_device_random_buffer(&avx_buf);
 
-    uint32_t crc = aws_checksums_crc32c_sw(avx_buf.buffer, avx_buf.len, 0);
-
-    uint32_t hw_crc = aws_checksums_crc32c_hw(avx_buf.buffer, avx_buf.len, 0);
+    uint32_t crc = aws_checksums_crc32c_sw(avx_buf.buffer, (int)avx_buf.len, 0);
+    uint32_t hw_crc = aws_checksums_crc32c_hw(avx_buf.buffer, (int)avx_buf.len, 0);
     ASSERT_UINT_EQUALS(hw_crc, crc);
 
     return res;
